requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
numpy
spacy
groq
easyocr
deep-translator
Pillow
langgraph
pytesseract
geopy
psycopg2-binary
